move state to model
