+++ src/morphometrics/_gui/_qt/labeling_widgets.py
@@ -12,15 +12,15 @@ from ...label.image_utils import (
     expand_selected_labels,
     get_mask_bounding_box_3d,
 )
+from ..models.labeling_model import LabelingModel
 
 
 class QtLabelingWidget(QWidget):
     def __init__(self, viewer: napari.Viewer):
         super().__init__()
         self._viewer = viewer
-        self._layer = None
-        self._background_mask_layer = None
-        self._curating = False
+        self._model = LabelingModel()
+        self._model.events.curating.connect(self._on_curating_change)
 
         # create the label selection widget
         self._label_selection_widget = magicgui(
@@ -43,78 +43,6 @@ class QtLabelingWidget(QWidget):
         self.layout().addWidget(self._label_selection_widget.native)
         self.layout().addWidget(self._label_expansion_widget.native)
 
-    @property
-    def layer(self) -> Optional[napari.layers.Labels]:
-        """The selected labels layer for curation.
-
-        Returns
-        -------
-        layer : Optional[napari.layers.Labels]
-            The selected labels layer that will be curated. If no layer is selected,
-            returns None.
-        """
-        return self._layer
-
-    @layer.setter
-    def layer(self, layer: Optional[napari.layers.Labels]):
-        """The selected labels layer for curation.
-
-        Parameters
-        ----------
-        layer : Optional[napari.layers.Labels]
-            The selected labels layer that will be curated. If no layer is selected,
-            set to None.
-        """
-        if layer is self._layer:
-            # if the layer hasn't changed, don't perform the update
-            return None
-        self._layer = layer
-        self._on_layer_update()
-
-    @property
-    def background_mask_layer(self) -> Optional[napari.layers.Labels]:
-        """The selected labels layer for curation.
-
-        Returns
-        -------
-        layer : Optional[napari.layers.Labels]
-            The selected labels layer that will be curated. If no layer is selected,
-            returns None.
-        """
-        return self._background_mask_layer
-
-    @background_mask_layer.setter
-    def background_mask_layer(self, layer: Optional[napari.layers.Labels]):
-        """The selected image layer for backgroundmaskiung.
-
-        Parameters
-        ----------
-        layer : Optional[napari.layers.Labels]
-            The selected image layer that be used to mask off background when expanding
-            labels
-        """
-        if layer is self._background_mask_layer:
-            # if the layer hasn't changed, don't perform the update
-            return None
-        self._background_mask_layer = layer
-        self._on_layer_update()
-
-    @property
-    def curating(self) -> bool:
-        return self._curating
-
-    @curating.setter
-    def curating(self, curating: bool):
-        if self._curating == curating:
-            return
-
-        if curating is True:
-            self._on_start_curating()
-        else:
-            self._on_stop_curating()
-
-        self._curating = curating
-
     def _expand_selected_labels_widget_function(
         self,
         pbar: widgets.ProgressBar,
@@ -125,10 +53,10 @@ class QtLabelingWidget(QWidget):
         label_values_to_expand = [int(value) for value in label_values_to_expand]
 
         # get the values from the selected labels layer
-        label_image = self.layer.data
-        label_layer_name = self.layer.name
-        if self.background_mask_layer is not None:
-            background_mask = self._background_mask_layer.data
+        label_image = self._model.labels_layer.data
+        label_layer_name = self._model.labels_layer.name
+        if self._model.background_mask_layer is not None:
+            background_mask = self._model.background_mask_layer.data
         else:
             background_mask = None
 
@@ -190,9 +118,9 @@ class QtLabelingWidget(QWidget):
         labels_layer: napari.layers.Labels,
         background_mask_layer: napari.layers.Image,
     ):
-        self.layer = labels_layer
-        self.background_mask_layer = background_mask_layer
-        self.curating = True
+        self._model.labels_layer = labels_layer
+        self._model.background_mask_layer = background_mask_layer
+        self._model.curating = True
 
     def _get_valid_labels_layers(self, combo_box) -> List[napari.layers.Labels]:
         return [
@@ -212,8 +140,8 @@ class QtLabelingWidget(QWidget):
 
         return valid_layers
 
-    def _on_start_curating(self):
-        self._label_expansion_widget.native.setVisible(True)
-
-    def _on_stop_curating(self):
-        self._label_expansion_widget.native.setVisible(False)
+    def _on_curating_change(self):
+        if self._model.curating is True:
+            self._label_expansion_widget.native.setVisible(True)
+        else:
+            self._label_expansion_widget.native.setVisible(False)
